ALIEN-462 : As a PaaS Provider developper I want to receive notifications when the list of compute template for a given cloud changes. ALIEN-458 : As a USER I want the cloud resource matching to be persisted in the deployment setup. ALIEN-459 : As a USER I want the cloud deployment properties to be persisted within the deployment setup ALIEN-460 : As a USER I want the cloud deployment properties to be persisted within the deployment setup

+++ alien4cloud-rest-api/src/main/java/alien4cloud/rest/cloud/CloudImageController.java
@@ -6,8 +6,6 @@ import java.util.UUID;
 import javax.annotation.Resource;
 import javax.validation.Valid;
 
-import lombok.extern.slf4j.Slf4j;
-
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -17,8 +15,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.google.common.collect.Sets;
+import com.wordnik.swagger.annotations.ApiOperation;
+import com.wordnik.swagger.annotations.ApiParam;
+import com.wordnik.swagger.annotations.Authorization;
+
 import alien4cloud.cloud.CloudImageService;
 import alien4cloud.dao.model.GetMultipleDataResult;
+import alien4cloud.exception.InvalidArgumentException;
 import alien4cloud.images.IImageDAO;
 import alien4cloud.images.exception.ImageUploadException;
 import alien4cloud.model.cloud.CloudImage;
@@ -26,10 +30,7 @@ import alien4cloud.model.cloud.CloudImageRequirement;
 import alien4cloud.rest.model.RestResponse;
 import alien4cloud.rest.model.RestResponseBuilder;
 import alien4cloud.utils.ReflectionUtil;
-
-import com.wordnik.swagger.annotations.ApiOperation;
-import com.wordnik.swagger.annotations.ApiParam;
-import com.wordnik.swagger.annotations.Authorization;
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * @author Minh Khang VU
@@ -50,7 +51,7 @@ public class CloudImageController {
      *
      * @param request The cloud image creation request.
      */
-    @ApiOperation(value = "Create a new cloud image.", authorizations = { @Authorization("ADMIN") })
+    @ApiOperation(value = "Create a new cloud image.", authorizations = {@Authorization("ADMIN")})
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public RestResponse<String> create(
             @ApiParam(value = "The instance of cloud image to add.", required = true) @Valid @RequestBody CloudImageCreateRequest request) {
@@ -69,19 +70,19 @@ public class CloudImageController {
         }
         cloudImageService.ensureCloudImageUniqueness(cloudImage);
         cloudImageService.saveCloudImage(cloudImage);
-        return RestResponseBuilder.<String> builder().data(cloudImage.getId()).build();
+        return RestResponseBuilder.<String>builder().data(cloudImage.getId()).build();
     }
 
     /**
      * Update an existing cloud image.
-     * 
-     * @param id id of the cloud image
+     *
+     * @param id      id of the cloud image
      * @param request the request which contains the cloud image details
      */
-    @ApiOperation(value = "Update an existing cloud image.", authorizations = { @Authorization("ADMIN") })
+    @ApiOperation(value = "Update an existing cloud image.", authorizations = {@Authorization("ADMIN")})
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public RestResponse<Void> update(@ApiParam(value = "The cloud image id to update.", required = true) @PathVariable String id,
-            @ApiParam(value = "The update request.", required = true) @Valid @RequestBody CloudImageUpdateRequest request) {
+                                     @ApiParam(value = "The update request.", required = true) @Valid @RequestBody CloudImageUpdateRequest request) {
         CloudImage cloudImage = cloudImageService.getCloudImageFailIfNotExist(id);
         // Register the state before the merge
         String oldName = cloudImage.getName();
@@ -92,7 +93,7 @@ public class CloudImageController {
             cloudImageService.ensureCloudImageUniqueness(cloudImage);
         }
         cloudImageService.saveCloudImage(cloudImage);
-        return RestResponseBuilder.<Void> builder().build();
+        return RestResponseBuilder.<Void>builder().build();
     }
 
     /**
@@ -100,11 +101,15 @@ public class CloudImageController {
      *
      * @param id Id of the cloud image to delete.
      */
-    @ApiOperation(value = "Delete an existing cloud image. The operation fails in case a cloud is still using the image.", authorizations = { @Authorization("ADMIN") })
+    @ApiOperation(value = "Delete an existing cloud image. The operation fails in case a cloud is still using the image.", authorizations = {@Authorization("ADMIN")})
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
     public RestResponse<Void> delete(@ApiParam(value = "Id of the cloud image to delete.", required = true) @PathVariable String id) {
+        String[] cloudsUsingImage = cloudImageService.getCloudsUsingImage(id);
+        if (cloudsUsingImage.length > 0) {
+            throw new InvalidArgumentException("Trying to delete a cloud image which is still used by following clouds " + Sets.newHashSet(cloudsUsingImage));
+        }
         cloudImageService.deleteCloudImage(id);
-        return RestResponseBuilder.<Void> builder().build();
+        return RestResponseBuilder.<Void>builder().build();
     }
 
     /**
@@ -112,10 +117,10 @@ public class CloudImageController {
      *
      * @param id Id of the cloud to delete.
      */
-    @ApiOperation(value = "Get details of a cloud image.", authorizations = { @Authorization("ADMIN") })
+    @ApiOperation(value = "Get details of a cloud image.", authorizations = {@Authorization("ADMIN")})
     @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     public RestResponse<CloudImage> get(@ApiParam(value = "Id of the cloud image for which to get details.", required = true) @PathVariable String id) {
-        return RestResponseBuilder.<CloudImage> builder().data(cloudImageService.getCloudImageFailIfNotExist(id)).build();
+        return RestResponseBuilder.<CloudImage>builder().data(cloudImageService.getCloudImageFailIfNotExist(id)).build();
     }
 
     /**
@@ -124,22 +129,22 @@ public class CloudImageController {
      * @param searchRequest Query to find the cloud images.
      * @return A {@link RestResponse} that contains a {@link alien4cloud.dao.model.GetMultipleDataResult} that contains the clouds.
      */
-    @ApiOperation(value = "Search for cloud images.", authorizations = { @Authorization("ADMIN") })
+    @ApiOperation(value = "Search for cloud images.", authorizations = {@Authorization("ADMIN")})
     @RequestMapping(value = "/search", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
     public RestResponse<GetMultipleDataResult> search(@RequestBody CloudImageSearchRequest searchRequest) {
         GetMultipleDataResult result = cloudImageService.get(searchRequest.getQuery(), searchRequest.getExclude(), searchRequest.getFrom(),
                 searchRequest.getSize());
-        return RestResponseBuilder.<GetMultipleDataResult> builder().data(result).build();
+        return RestResponseBuilder.<GetMultipleDataResult>builder().data(result).build();
     }
 
     /**
      * Update cloud image's icon.
      *
-     * @param id The cloud image id.
+     * @param id    The cloud image id.
      * @param image new icon of the cloud image.
      * @return nothing if success, error will be handled in global exception strategy
      */
-    @ApiOperation(value = "Updates the icon for the cloud image.", authorizations = { @Authorization("ADMIN") })
+    @ApiOperation(value = "Updates the icon for the cloud image.", authorizations = {@Authorization("ADMIN")})
     @RequestMapping(value = "/{id}/icon", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
     public RestResponse<String> updateImage(@PathVariable String id, @RequestParam("file") MultipartFile image) {
         CloudImage cloudImage = cloudImageService.getCloudImageFailIfNotExist(id);
@@ -151,6 +156,6 @@ public class CloudImageController {
         }
         cloudImage.setIconId(iconId);
         cloudImageService.saveCloudImage(cloudImage);
-        return RestResponseBuilder.<String> builder().data(iconId).build();
+        return RestResponseBuilder.<String>builder().data(iconId).build();
     }
 }
